refactor: 클래스형 문법에서 함수형 문법으로 이전
